Busybox resize command
We can force the terminal to run busybox resize.  I think this will reduce errors such as:
```
root@bacon:/ # 1;148R
```
because the app is running vice the bootkali script, which is where most of the errors seem to be coming form.

This requires busybox to be in the path…however if the user doesn’t have busybox then they can’t chroot anyways.

Signed-off-by: binkybear <binkybear@nethunter.com>